scripts/requirements.txt
Python-SimConnect==0.4.26
